Updated squadmortar Overlay functions
