+++ public/locales/uk/tooltips.json
@@ -6,6 +6,7 @@
     "focusMode": "Режим фокусу",
     "enterToExit": "Натисніть 'Enter', щоб вийти",
     "squadMortarOverlay": "SquadMortarOverlay detected",
+    "Карту оновлено": "Карту оновлено",
     "connectedTo": "Connected !",
     "error": "Помилка!",
     "apiError": "API не підключено..",
