+++ src/js/squadCalc.js
@@ -7,7 +7,8 @@ import { loadSettings } from "./settings.js";
 import { loadLanguage } from "./localization.js";
 import { animateCSS, animateCalc } from "./animations.js";
 import { tooltip_save, tooltip_copied } from "./tooltips.js";
-import { checkApiHealth, initWebSocket } from "./squadCalcAPI.js";
+import { checkApiHealth } from "./squadCalcAPI.js";
+import { initWebSocket } from "./smcConnector.js";
 import SquadFiringSolution from "./squadFiringSolution.js";
 import packageInfo from "../../package.json";
 import i18next from "i18next";
