+++ src/js/squadCalcAPI.js
@@ -1,6 +1,4 @@
 import packageInfo from "../../package.json";
-import { App } from "../app.js";
-
 
 /**
  * Sends marker data to the API via a POST request.
@@ -97,134 +95,4 @@ export async function fetchMarkersByMap(mapName, weapon) {
         console.debug("Error fetching marker data:", error);
         throw error;
     }
-}
-
-
-
-export async function initWebSocket() {
-
-    // Disable WebSockets if not activated
-    if (process.env.WEBSOCKET != "true") { return; }
-    if (!await checkServerAvailability()) { return; }
-
-    const socketMap = new WebSocket("ws://127.0.0.1:12345");
-    const socketCoordinates = new WebSocket("ws://127.0.0.1:12346");
-    setInterval(() => checkCoordinates(socketCoordinates), 1000);
-    
-    socketMap.addEventListener("open", () => {
-        console.log("Connected to SquadMortarOverlay.exe");
-        App.openToast("success", "connectedTo", "squadMortarOverlay");
-    });
-    
-    socketMap.addEventListener("message", async (event) => {
-        if (event.data === "Map") {
-            if (socketMap.readyState === WebSocket.OPEN) {
-                let imageUrl = "maps" + App.minimap.activeMap.mapURL + "basemap.webp";
-                // Fetch the image and send its binary data
-                const response = await fetch(imageUrl);
-                const imageBlob = await response.blob();
-                const reader = new FileReader();
-                reader.onload = function() {
-                    const arrayBuffer = this.result;
-                    socketMap.send(arrayBuffer);
-                };
-                reader.readAsArrayBuffer(imageBlob);
-                App.minimap.changeLayer();
-            }
-        }
-    });
-    
-    socketMap.addEventListener("message", (event) => {
-        if (event.data instanceof Blob) {
-            const url = URL.createObjectURL(event.data);
-            App.minimap.activeLayer.setUrl(url);
-            const notification = document.createElement("div");
-            notification.innerText = "Map Updated!";
-            notification.style.position = "fixed";
-            notification.style.top = "15px";
-            notification.style.left = "50%";
-            notification.style.transform = "translateX(-50%)";
-            notification.style.backgroundColor = "green";
-            notification.style.padding = "10px";
-            notification.style.borderRadius = "5px";
-            notification.style.zIndex = "1000";
-            notification.style.opacity = "0";
-            notification.style.transition = "opacity 0.5s";
-            document.body.appendChild(notification);
-            
-            // Fade in
-            setTimeout(() => {
-                notification.style.opacity = "1";
-            }, 10); // Delay to ensure the element is rendered before starting the fade-in
-            
-            // Fade out after 3 seconds
-            setTimeout(() => {
-                notification.style.opacity = "0";
-                // Remove the notification after fade out
-                setTimeout(() => {
-                    document.body.removeChild(notification);
-                }, 500); // Match the duration of the opacity transition
-            }, 2500);
-        }
-    });
-}
-
-
-function checkCoordinates(socketCoordinates) {
-    let coorArray = "";
-    let prevCoorArray = "";
-    if (App.minimap.activeTargetsMarkers) {
-        App.minimap.activeTargetsMarkers.eachLayer(function (target) {
-            const BEARING = target.firingSolution1.bearing;
-            let elevation;
-            if (App.minimap.activeWeaponsMarkers.getLayers()[0].angleType === "high"){
-                elevation = target.firingSolution1.elevation.high;
-            }
-            else {
-                elevation = target.firingSolution1.elevation.low;
-            }
-
-            if (isNaN(elevation)) {
-                elevation = "---";
-            } else {
-                if (App.activeWeapon.unit === "mil") {
-                    elevation = elevation.mil.toFixed(0);
-                } else {
-                    elevation = elevation.deg.toFixed(1);
-                }
-            }
-            if (coorArray !== "") {
-                coorArray = coorArray + "\n";
-            }
-            coorArray = coorArray + `${elevation} | ${BEARING.toFixed(1)}°`;
-        });
-    }
-    if (coorArray !== prevCoorArray) {
-        prevCoorArray = coorArray;
-        if (socketCoordinates.readyState === WebSocket.OPEN) {
-            socketCoordinates.send(coorArray);
-        }
-    }
-}
-
-/**
- * Checks the availability of a SquadMortarOverlay.exe by attempting to load a fake image from it
- * @param {number} [timeout] - The maximum time to wait for response
- * @returns {Promise<boolean>} A promise that resolves to `true` if the SquadMortarOverlay.exe is running
- */
-function checkServerAvailability(timeout = 2000) {
-    return new Promise((resolve) => {
-        const img = new Image();
-        const timer = setTimeout(() => {
-            img.src = "";
-            resolve(false);
-        }, timeout);
-
-        img.onload = img.onerror = () => {
-            clearTimeout(timer);
-            resolve(true);
-        };
-        
-        img.src = "http://127.0.0.1:12345";
-    });
 }
